Delete GIFTs client
The client doesn't work as is, but actually, we don't need to maintain it - there's no use case for it. The GIFTs service needs to be used maybe 10 times a year, so we don't need any sort of scripted bulk submission, and the GUI is going to be the easiest way to do things.