Remove name_matches and excluded_schemas parameters from get_database_set
Rewrite the signature to simplify the behaviour.
Replace with incl_filters and skip_filters instead.

The behaviour is simplified by introducting incl_filters and
skip_filters which will respectively accept a list of patterns used to
filter out names that aren't matched, and a list of patterns used to
filter out names that are matched.